refactor(frontend): move bubble and textarea to the chat folder
