--- frontend/apps/web/app/chat-example/page.tsx
@@ -1,1 +1,0 @@
-
