+++ frontend/apps/web/components/chat/bubble.tsx
@@ -11,20 +11,20 @@ interface ChatBubbleProps {
 
 export function ChatBubble({ content, isUser = true, className }: ChatBubbleProps) {
   return (
-    <div className="w-full px-2 sm:px-4 py-2">
+    <div className="w-full px-2 sm:px-4 py-1 sm:py-2">
       <div className="max-w-3xl mx-auto">
         <div 
           className={cn(
-            "overflow-hidden p-3 sm:p-4",
+            "rounded-xl overflow-hidden p-2 sm:p-3 md:p-4",
             isUser 
-              ? "bg-lime-50 rounded-xl dark:bg-lime-950/50 border-lime-200 dark:border-lime-800/30 border max-w-lg ml-auto" 
-              : "w-full",
+              ? "bg-lime-50 dark:bg-lime-950/50 border-lime-200 dark:border-lime-800/30 border max-w-[80%] sm:max-w-[70%] md:max-w-lg ml-auto" 
+              : "w-full bg-sidebar",
             className
           )}
         >
-          <div className="flex flex-col gap-2 sm:gap-3">
+          <div className="flex flex-col gap-1 sm:gap-2">
             <div className="relative w-full">
-              <p className="w-full text-base md:text-sm px-3 py-2">
+              <p className="w-full text-base md:text-sm px-2 sm:px-3 py-1 sm:py-2">
                 {content}
               </p>
             </div>
