+++ frontend/apps/web/components/chat/textarea.tsx
